fixed the ut
